% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_tools.R
\name{compute_distance}
\alias{compute_distance}
\title{compute_distance - a function for computing similarities on elements of a vector space, as in L1000 data}
\usage{
compute_distance(
  ds,
  gene_annot,
  upgenes = c(),
  dngenes = c(),
  geneweights = 0,
  geneset_name = "",
  metric = "cmap_score",
  gseaParam = 1,
  nperms = 1000
)
}
\arguments{
\item{ds}{}

\item{gene_annot}{}

\item{upgenes}{}

\item{dngenes}{}

\item{geneweights}{}

\item{geneset_name}{}

\item{metric}{}

\item{gseaParam}{}

\item{nperms}{}
}
\value{
data frame
}
\description{
compute_distance - a function for computing similarities on elements of a vector space, as in L1000 data
}
