% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sim_block.R
\name{compute_sim_block}
\alias{compute_sim_block}
\title{compute_sim_block}
\usage{
compute_sim_block(
  ds1,
  ds2,
  metric = "fastcosine",
  kgenes = 0,
  gseaParam = 1,
  nperms = 1,
  parallel = 0,
  numCores = parallel::detectCores()
)
}
\arguments{
\item{ds1}{A gct instance, dataset 1}

\item{ds2}{A gct instance, dataset 2 (could be the same as dataset 1)}

\item{metric}{A string, one of wtcs, cosine, spearman, pearson}

\item{kgenes}{For those metrics that restrict the number of genes (wtcs, cosine), a parameter that determines how many genes are used to compute similarity. Default is 50 for GSEA, 0 (all) for cosine.}

\item{gseaParam}{The GSEA parameter that affects the weighting of the gene values in GSEA.  Default is 1.}

\item{nperms}{Numeric, indicates how many permutations are run by fgsea for significance calculations.  As p-values from GSEA are not used, this defaults to 1.}
}
\value{
matrix
}
\description{
This function takes as input two gct datasets and computes a similarity matrix between
columns of the first dataset and columns of the second.  It requires the gene space to be
common to both.  Note that some similarity metrics, like GSEA-based metrics, are not symmetric.
}
