% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine.R
\name{cosine}
\alias{cosine}
\title{cosine}
\usage{
cosine(mat1, mat2)
}
\arguments{
\item{mat1}{Numeric, a matrix of size N x d1}

\item{mat2}{Numeric, a matrix of size N x d2}
}
\value{
matrix
}
\description{
This function computes the cosine distance between all pairs of columns of two input matrices
}
