% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{download_l1k_data}
\alias{download_l1k_data}
\title{download_l1k_data - download data for L1000 from GEO and clue.io}
\usage{
download_l1k_data(mypath, version = 2017)
}
\arguments{
\item{mypath}{Path to the file to which the downloaded data is saved}

\item{version}{numeric, either 2017 or 2020}
}
\description{
download_l1k_data - download data for L1000 from GEO and clue.io
}
