% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fgseaBase.R
\name{fgseaBase}
\alias{fgseaBase}
\title{fgseaBase}
\usage{
fgseaBase(
  pathways,
  stats,
  nperm,
  minSize = 1,
  maxSize = Inf,
  scoreType = c("std", "pos", "neg"),
  nproc = 0,
  gseaParam = 1,
  BPPARAM = NULL
)
}
\arguments{
\item{pathways}{List of gene sets to check}

\item{stats}{Named vector of gene-level stats. Names should be the same as in 'pathways'}

\item{nperm}{Number of permutations to do. Minimial possible nominal p-value is about 1/nperm}

\item{minSize}{Minimal size of a gene set to test. All pathways below the threshold are excluded.}

\item{maxSize}{Maximal size of a gene set to test. All pathways above the threshold are excluded.}

\item{scoreType}{This parameter defines the GSEA score type. Possible options are ("std", "pos", "neg")}

\item{nproc}{If not equal to zero sets BPPARAM to use nproc workers (default = 0).}

\item{gseaParam}{GSEA parameter value, all gene-level statis are raised to the power of 'gseaParam' before calculation of GSEA enrichment scores.}

\item{BPPARAM}{Parallelization parameter used in bplapply. Can be used to specify cluster to run. If not initialized explicitly or by setting 'nproc' default value 'bpparam()' is used.}
}
\value{
A table with GSEA results. Each row corresponds to a tested pathway. The columns are:
\itemize{
\item pathway -- name of the pathway
\item ES -- enrichment score, the GSEA test statistic
\item NES -- normalized enrichment score, the ES divided by the mean value for pathways of the same size. An NES of 1 \link{or -1} corresponds to the centre of the null distribution (for positively \link{or negatively} scored pathways)
\item size -- number of genes in the pathway after removing genes not present in names(stats)
}
}
\description{
This function is a modification of the fgsea function that dispenses with the adaptive permutation testing
and doesn't return or calculate leading edge genes. This makes it much faster when the only objective is
calculation of the raw GSEA enrichment score (e.g. for use as a distance function).
}
