% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get_level5_ds}
\alias{get_level5_ds}
\title{get_level5_ds - return the path to the level 5 dataset in a directory to simplify grabbing file names.}
\usage{
get_level5_ds(mypath, mypattern = "")
}
\arguments{
\item{mypath}{Path to the file to which the downloaded data is saved}
}
\value{

}
\description{
get_level5_ds - return the path to the level 5 dataset in a directory to simplify grabbing file names.
}
