% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopK.R
\name{getTopK}
\alias{getTopK}
\title{getTopK}
\usage{
getTopK(dsmat, k = 50, decreasing = TRUE)
}
\arguments{
\item{dsmat}{A matrix}

\item{k}{A non-negative integer indicating the number of largest elements to return}

\item{decreasing}{A boolean indicating whether to get the largest elements (TRUE) or the smallest elements (FALSE); true by default.}
}
\value{
vector
}
\description{
This is a helper function used in compute_sim_block that identifies the row identifiers of the K largest or smallest elements in each column of a matrix.
}
