% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{read_l1k_meta}
\alias{read_l1k_meta}
\title{read_l1k_meta - load metadata for L1000 datasets}
\usage{
read_l1k_meta(mypath, version = 2017)
}
\arguments{
\item{mypath}{Path to data}

\item{version}{numeric, either 2017 or 2020}
}
\value{
list
}
\description{
read_l1k_meta - load metadata for L1000 datasets
}
