% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryBenchmark.R
\name{queryBenchmark}
\alias{queryBenchmark}
\title{queryBenchmark - Benchmark the performance of L1000 queries}
\usage{
queryBenchmark(ds, parallel = 1, k = c(10, 30, 100, 300))
}
\arguments{
\item{ds}{The dataset - a  GCT object (e.g. a gctx file from cmapR). The assumed structure has columns are samples, rows are genes.}

\item{parallel}{Boolean, whether to evaluate the multicore performance from the library parallel.}

\item{k}{Numeric vector of values to evaluate, typically 10, 100, 1000}
}
\value{

}
\description{
queryBenchmark - Benchmark the performance of L1000 queries
}
