% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_l1k_meta.R
\name{read_l1k_meta}
\alias{read_l1k_meta}
\title{read_l1k_meta}
\usage{
read_l1k_meta(mypath, version = 2017)
}
\arguments{
\item{mypath}{A string with the path to the directory containing the metadata, with subfolders 2017 and 2020.}

\item{version}{A numeric indicating which version of the data to use - 2017 or 2020 (defaults to 2017).}
}
\description{
This function parses the L1000 metadata downloaded from either GEO 92742 or the clue.io data repository,
parses it, and returns the meta data as a list of dataframes.
}
